+++ examples/advanced-blog/src/lib/mdx-server.ts
@@ -8,29 +8,26 @@ export default async function MDXServer(code: string) {
   const result = await bundleMDX({
     source: code,
     mdxOptions(options) {
-      ;(options.remarkPlugins = [
-        ...(options.remarkPlugins ?? []),
-        remarkGfm // Adding remarkGfm here
-      ]),
-        (options.rehypePlugins = [
-          ...(options.rehypePlugins ?? []),
-          rehypeSlug,
-          [
-            rehypePrettyCode,
-            {
-              theme: 'dracula'
-              // The rest of the rehypePrettyCode config
-            }
-          ],
-          [
-            rehypeAutolinkHeadings,
-            {
-              properties: {
-                className: ['hash-anchor']
-              }
-            }
-          ]
-        ])
+      options.remarkPlugins = options.remarkPlugins ?? []
+      options.remarkPlugins.push(remarkGfm as any)
+
+      options.rehypePlugins = options.rehypePlugins ?? []
+      options.rehypePlugins.push(rehypeSlug as any)
+      options.rehypePlugins.push([
+        rehypePrettyCode as any,
+        {
+          theme: 'dracula' // Add other rehypePrettyCode options if necessary
+        }
+      ])
+      options.rehypePlugins.push([
+        rehypeAutolinkHeadings as any,
+        {
+          properties: {
+            className: ['hash-anchor']
+          }
+        }
+      ])
+
       return options
     }
   })