+++ packages/outstatic/src/components/DocumentSettings/index.tsx
@@ -16,6 +16,7 @@ import { useRouter } from 'next/navigation'
 import { useContext, useState } from 'react'
 import { RegisterOptions, useFormContext } from 'react-hook-form'
 import { slugify } from 'transliteration'
+import { CheckboxWithLabel } from '@/components/ui/outstatic/checkbox-with-label'
 import { Button } from '@/components/ui/shadcn/button'
 import useOutstatic from '@/utils/hooks/useOutstatic'
 
@@ -28,13 +29,17 @@ type DocumentSettingsProps = {
 }
 
 interface CustomInputProps {
-  type?: 'text' | 'number'
+  type?: 'text' | 'number' | 'checkbox'
   suggestions?: CustomFieldArrayValue[]
   registerOptions?: RegisterOptions
 }
 
 type ComponentType = {
-  component: typeof Input | typeof TextArea | typeof TagInput
+  component:
+    | typeof Input
+    | typeof TextArea
+    | typeof TagInput
+    | typeof CheckboxWithLabel
   props: CustomInputProps
 }
 
@@ -43,6 +48,7 @@ type FieldDataMapType = {
   Text: ComponentType
   Number: ComponentType
   Tags: ComponentType
+  Boolean: ComponentType
 }
 
 const FieldDataMap: FieldDataMapType = {
@@ -54,7 +60,8 @@ const FieldDataMap: FieldDataMapType = {
     props: {
       suggestions: []
     }
-  }
+  },
+  Boolean: { component: CheckboxWithLabel, props: { type: 'checkbox' } }
 }
 
 const DocumentSettings = ({